+++ _sql/v2.sql
@@ -4,3 +4,19 @@ ALTER TABLE eshop_product DROP COLUMN unavailable;
 ALTER TABLE eshop_product DROP COLUMN priority;
 ALTER TABLE eshop_product DROP COLUMN recommended;
 ALTER TABLE eshop_product DROP COLUMN primaryCategory;
+
+DROP TABLE eshop_parametervalue;
+DROP TABLE eshop_parameteravailablevalue;
+DROP TABLE eshop_parametergroup;
+DROP TABLE eshop_parametercategory;
+DROP TABLE eshop_parameter;
+
+DROP TABLE eshop_producttabtext;
+DROP TABLE eshop_producttab;
+
+DROP TABLE eshop_setitem;
+DROP TABLE eshop_set;
+
+DROP TABLE eshop_product_nxn_eshop_tag;
+DROP TABLE eshop_tag_nxn_eshop_tag;
+DROP TABLE eshop_tag;