+++ src/Admin/ProductPresenter.php
@@ -32,9 +32,8 @@ use Eshop\DB\PricelistRepository;
 use Eshop\DB\PriceRepository;
 use Eshop\DB\ProducerRepository;
 use Eshop\DB\Product;
+use Eshop\DB\ProductContentRepository;
 use Eshop\DB\ProductRepository;
-use Eshop\DB\ProductTabRepository;
-use Eshop\DB\ProductTabTextRepository;
 use Eshop\DB\RelatedTypeRepository;
 use Eshop\DB\StoreRepository;
 use Eshop\DB\SupplierProductRepository;
@@ -129,7 +128,6 @@ class ProductPresenter extends BackendPresenter
 		],
 		'detailSuppliersTab' => false,
 		'extendedName' => false,
-		'productTabs' => true,
 		'karsa' => false,
 		ProductFormConfig::class => [
 			ProductFormAutoPriceConfig::class => ProductFormAutoPriceConfig::NONE,
@@ -163,10 +161,7 @@ class ProductPresenter extends BackendPresenter
 	public PriceRepository $priceRepository;
 
 	#[\Nette\DI\Attributes\Inject]
-	public ProductTabRepository $productTabRepository;
-
-	#[\Nette\DI\Attributes\Inject]
-	public ProductTabTextRepository $productTabTextRepository;
+	public ProductContentRepository $productContentRepository;
 
 	#[\Nette\DI\Attributes\Inject]
 	public VatRateRepository $vatRateRepository;
@@ -493,32 +488,6 @@ class ProductPresenter extends BackendPresenter
 		 */
 		foreach ($input->getComponents() as $pricelistId => $container) {
 			$container->setDefaults($prices[$pricelistId]->toArray());
-		}
-
-		/** @var \Eshop\DB\ProductTab $productTab */
-		foreach ($this->productTabRepository->many() as $productTab) {
-			$values = [];
-			$productTabText = $this->productTabTextRepository->many()
-				->where('fk_product=:product AND fk_tab=:tab', ['product' => $product->getPK(), 'tab' => $productTab->getPK()])
-				->first();
-
-			if (!$productTabText) {
-				continue;
-			}
-
-			/** @var \Forms\Container $input */
-			$input = $form['productTab' . $productTab->getPK()];
-
-			/**
-			 * @var string $key
-			 * @var \Forms\Container $container
-			 */
-			foreach ($input->getComponents() as $key => $container) {
-				$property = 'content_' . $key;
-				$values[$key] = $productTabText->$property;
-			}
-
-			$input->setDefaults($values);
 		}
 
 		$amounts = $this->amountRepository->many()